Add a NIPSA service to replace h.nipsa.logic
One approach to limiting coupling between views and the models they
control is to introduce a service layer which is responsible for
encapsulating business logic and dependencies of that logic on other
parts of the application.

This commit uses the `pyramid_services` library, which is a small
library formalising a service layer pattern, in which services are
registered with the application registry by name or by interface (I've
chosen to go with name here, for the sake of simplicity). These services
can then be retrieved by anything which can call
`request.find_service(...)`.

Crucially, one can register service factories which in turn have access
to the request object and thus can extract other dependencies (such as
the database session) without calling code having to pass them in. This
is the key difference between this approach and the more direct "logic"
module approach.